Decided to implement OTP for email. Removed (auth) forgot password logic and confirmationUrl signUp flow. Implemented successfully although couldn't redirect to /complete-registration route whilst route was inside of (auth) route group for some reason so had to take it out. Images load no issues. Re-styled Login and Signup pages and added both of their verify-otp routes. Removed icons, link for security route and deleted /security route in LayoutNav. Replace all instances of hint accentRed with deepOlive. Constructed post request and created new api route endpoint that checks profiles table for existing email during signup to prevent users signing in through the signup process. In order to do this check i needed to bypass rls by using my supabase service role key. Implemented user feedback and redirect to login for existing accounts. Tested. All working.

+++ app/(auth)/signup/page.jsx
@@ -17,6 +17,7 @@ const Signup = () => {
   const [isLoading, setIsLoading] = useState(false)
   const [isChecked, setIsChecked] = useState(false)
   const [error, setError] = useState(null)
+  const [checkBoxError, setCheckBoxError] = useState(null)
   const router = useRouter()
 
 
@@ -30,57 +31,101 @@ const Signup = () => {
     setIsChecked(e.target.checked)
   }
 
+
+
+
+
+
+
+
   const handleSubmit = async (e) => {
     e.preventDefault()
 
     // form validation
     if (!email) {
-        setError('Please provide your email');
-        setTimeout(() => setError(null), 2000)
-        setIsLoading(false)
-        return
-
+      setError('Please provide your email');
+      setTimeout(() => setError(null), 2000);
+      setIsLoading(false);
+      return;
     } else if (!isValidEmail(email)) {
-        setError('Unable to validate email address: invalid format');
-        setTimeout(() => setError(null), 2000)
-        setIsLoading(false)
-        return
-
+      setError('Unable to validate email address: invalid format');
+      setTimeout(() => setError(null), 2000);
+      setIsLoading(false);
+      return;
     } else if (!isChecked) {
-        setError('Please confirm you have agreed to the Privacy Policy and Terms of Service');
-        setTimeout(() => setError(null), 2000)
-        setIsLoading(false)
-        return
+      setCheckBoxError(
+        'Please confirm you have agreed to the privacy policy and terms of service.'
+      );
+      setTimeout(() => setCheckBoxError(null), 2000);
+      setIsLoading(false);
+      return;
     }
 
     setIsLoading(true)
 
-    // store email temporarily in local storage
-    localStorage.setItem('email', email)
+
 
     
-    const supabase = createClientComponentClient()
-    const { data, error } = await supabase.auth.signInWithOtp({
-      email
-    })
+    // sent email to server endpoint to check if email already exists within profiles table
+    try {
+        const res = await fetch(`${location.origin}/api/auth/email-exists`, {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({
+            email
+          })
+        })
+
+        const userEmail = await res.json()
+        console.log('Server response:', userEmail);
+
+
+        if (userEmail.exists) {
+          setIsLoading(false)
+          setError('This email is already associated with an account. Please login.')
+          return
+        } else {
+            console.log('the email does not exist!')
     
+            // store email temporarily in local storage
+            localStorage.setItem('email', email)
 
-    if (error) {
-      setError(error.message)
-      setTimeout(() => setError(null), 2000)
-      setIsLoading(false);
-    } 
+            
+            const supabase = createClientComponentClient()
+            const { data, error } = await supabase.auth.signInWithOtp({
+              email
+            })
+            
 
-    if (!error) {
-      router.push('/verify-signup-otp')
+            if (error) {
+              setError(error.message)
+              setTimeout(() => setError(null), 2000)
+              setIsLoading(false);
+            } 
+
+            if (!error) {
+              router.push('/verify-signup-otp')
+            }
+
+        }
+
+    } catch (error) {
+        console.log(error)
     }
+
   }
 
+
+
+
+
     return (
-      <div className='flex flex-col items-center gap-12 md:flex-row md:justify-evenly md:gap-0 mb-4.5 md:flex-row md:h-auth-page-height'>
+      <div className='flex flex-col items-center gap-12 mb-4.5 md:justify-evenly md:gap-0 md:flex-row md:h-auth-page-height'>
 
         <form onSubmit={handleSubmit} className="w-full max-w-xs">
-          <h2 className='text-3xl mb-6 font-eb text-accentRed'>Sign up</h2>
+          <h2 className='text-3xl mb-6 font-eb text-deepOlive'>Sign up</h2>
           <p className='mb-3'>Enter your email to recieve a OTP (One-Time Passcode) to create your account.</p>
           
           <label>
@@ -88,7 +133,7 @@ const Signup = () => {
               Email
             </span>
             <input
-              className="w-full p-2.5 rounded-md text-stoneGray shadow-inner bg-deepCharcoal border-2 border-stoneGray"
+              className={`w-full p-2.5 rounded-md text-stoneGray shadow-inner bg-nightSky border-2 ${error ? 'border-red-900' : 'border-stoneGray'}`}
               type='text'
               spellCheck='false'
               value={email}
@@ -98,12 +143,16 @@ const Signup = () => {
 
           <div className="mt-5 flex items-center">
             <input className="self-start mt-0.21 max-w-min transform scale-125" type="checkbox" value={isChecked} onChange={handleCheckbox}/>
-            <span className="text-base block text-stoneGray ml-2.5 -mt-1">By signing up I agree to the{' '}<Link className="text-accentRed text-base" href='#'>Privacy Policy</Link>{' '}and{' '}<Link className='text-accentRed text-base' href='#'>Terms of Service</Link>
+            <span className="text-base block text-stoneGray ml-2.5 -mt-1">By signing up I agree to the{' '}<Link className="text-deepOlive text-base" href='#'>Privacy Policy</Link>{' '}and{' '}<Link className='text-deepOlive text-base' href='#'>Terms of Service</Link>
             .</span>
           </div>
 
-          <button className='btn block mt-4 bg-accentRed'>{isLoading ? 'Processing...' : 'Signup'}</button>
-          {error && <div className="mt-4 text-center error">{error}</div>}
+          <button className='btn block mt-4 bg-deepOlive'>{isLoading ? 'Processing...' : 'Signup'}</button>
+          <div className="mt-5 h-5 text-center">
+            {error && <div className="error">{error}</div>}
+            {checkBoxError && <div className="error">{checkBoxError}</div>}
+          </div>
+
         </form>
         
 
@@ -112,7 +161,7 @@ const Signup = () => {
           <SocialButtons text={"Continue"} />
           <div className="mt-7">
             <p className='inline mt-8 pr-2'>Have an account?</p>
-            <Link className='text-base text-accentRed' href='/login'>Login</Link>
+            <Link className='text-base text-deepOlive' href='/login'>Login</Link>
           </div>
         </div>
